START OF THE REBOOT!
After many erros and different aproaches, the best option is to migrate from maven to gradle and also merge and refactor all the code from other branches, and also eliminate anything related to netty, to many errors and pain with that dependency with lambdas. Many commits are not public as there are some secrets that end up being commited.
Local development will use sam and tests containers, using adapters and other technologies to mimic AWS, specially DynamoDB, are a bad solution, as many AWS related erros can not be replicated with other tools and viceversa.

The changes are going to be gradual and will be commited.

Changes:
- Deleted Core